clean architecture and use datalist
