+++ assets/scripts/inputs/inputMain.js
@@ -2,7 +2,7 @@ import {
   updateItems,
   updateRecipesFromMainInput,
   updateRecipesToShowAll,
-} from '../data/updateData.js';
+} from '../recipes/updateRecipes.js';
 
 export const inputMain = (e) => {
   const input = e.target.value;
@@ -16,4 +16,4 @@ export const inputMain = (e) => {
   updateRecipesToShowAll();
 };
 
-export const inputSecondary = () => {};
+export default inputMain;
