+++ assets/scripts/recipes/showRecipes.js
@@ -1,39 +1,5 @@
 import recipes from '../../data/recipes.js';
 import domElements from '../domElements.js';
-import { getData } from './cleanData.js';
-
-const { allIngredients, allAppliances, allUstensils } = getData();
-
-export const showInputsSecondary = () => {
-  domElements.inputSecondary.forEach((input) => {
-    const { lastElementChild: inputLEC } = input;
-    const inputSecondaryName = input.classList[1];
-    let array = [];
-    if (inputSecondaryName === 'ingredients') {
-      array = allIngredients;
-    } else if (inputSecondaryName === 'appliances') {
-      array = allAppliances;
-    } else if (inputSecondaryName === 'ustensils') {
-      array = allUstensils;
-    }
-
-    inputLEC.replaceChildren();
-
-    array
-      .sort((a, b) => a.recurrence < b.recurrence)
-      .forEach((item, index) => {
-        const div = document.createElement('div');
-        div.textContent =
-          item.item.substring(0, 1).toUpperCase() + item.item.substring(1);
-        inputLEC.appendChild(div);
-        div.className = 'input-secondary-ingredient';
-
-        if (index < 30) {
-          div.classList.add('show');
-        }
-      });
-  });
-};
 
 export const showRecipes = () => {
   recipes.forEach((recipe) => {
@@ -105,3 +71,5 @@ export const showRecipes = () => {
     domElements.allRecipes.appendChild(article);
   });
 };
+
+export default showRecipes;
