+++ assets/scripts/recipes/updateRecipes.js
@@ -1,5 +1,6 @@
 import recipes from '../../data/recipes.js';
 import domElements from '../domElements.js';
+import { showElement, hideElement } from '../tools/element.js';
 
 const isIncludesInName = (recipe, input) =>
   recipe.name.toLowerCase().includes(input.toLowerCase());
@@ -21,27 +22,14 @@ const isIncludesInUstensils = (recipe, input) => {
   );
 };
 
-const showElement = (element) => {
-  if (!element.classList.contains('show')) {
-    element.classList.add('show');
-  }
-};
-
-const hideElement = (element) => {
-  if (element.classList.contains('show')) {
-    element.classList.remove('show');
-  }
-};
-
 export const updateRecipesFromMainInput = (input) => {
   const allDomRecipes = document.querySelectorAll('.recipe');
   const recipesUpdate = [];
-  const DOMFields = domElements.fields;
-  const fields = [];
-  DOMFields.childNodes.forEach((child) => {
-    fields.push({ item: child.classList[1], name: child.textContent });
+  const DOMTags = domElements.tags;
+  const tags = [];
+  DOMTags.childNodes.forEach((child) => {
+    tags.push({ item: child.classList[1], name: child.textContent });
   });
-  console.log(fields);
 
   recipes.forEach((recipe) => {
     const thisDomRecipe = Array.from(allDomRecipes).find(
@@ -54,39 +42,39 @@ export const updateRecipesFromMainInput = (input) => {
       isIncludesInName(recipe, input) ||
       isIncludesInUstensils(recipe, input);
 
-    const isFieldIncludesInRecipeFunction = () => {
-      let checked = false;
-      if (fields.length === 0) {
-        checked = true;
-        return checked;
-      }
-      fields.forEach((field) => {
-        if (field.item === 'appliances') {
-          if (isIncludesInAppliances(recipe, field.name)) {
-            checked = true;
-            return;
-          }
-        }
-        if (field.item === 'ustensils') {
-          if (isIncludesInUstensils(recipe, field.name)) {
-            checked = true;
-            return;
-          }
-        }
-        if (field.item === 'ingredients') {
-          if (isIncludesInIngredients(recipe, field.name)) {
-            checked = true;
-          }
-        }
-      });
+    // const isTagIncludesInRecipeFunction = () => {
+    //   let checked = false;
+    //   if (tags.length === 0) {
+    //     checked = true;
+    //     return checked;
+    //   }
+    //   tags.forEach((tag) => {
+    //     if (tag.item === 'appliances') {
+    //       if (isIncludesInAppliances(recipe, tag.name)) {
+    //         checked = true;
+    //         return;
+    //       }
+    //     }
+    //     if (tag.item === 'ustensils') {
+    //       if (isIncludesInUstensils(recipe, tag.name)) {
+    //         checked = true;
+    //         return;
+    //       }
+    //     }
+    //     if (tag.item === 'ingredients') {
+    //       if (isIncludesInIngredients(recipe, tag.name)) {
+    //         checked = true;
+    //       }
+    //     }
+    //   });
 
-      console.log(checked);
-      return checked;
-    };
+    //   console.log(checked);
+    //   return checked;
+    // };
 
-    // const isFieldIncludesInRecipe = isFieldIncludesInRecipeFunction();
+    // const isTagIncludesInRecipe = isTagIncludesInRecipeFunction();
 
-    if (isInputIsIncludesInRecipe && isFieldIncludesInRecipeFunction()) {
+    if (isInputIsIncludesInRecipe) {
       showElement(thisDomRecipe);
       recipesUpdate.push(recipe.id);
       return;
